requirements.txt
fastmcp
openai
datadog-api-client
PyGithub
jira
requests
psycopg2-binary
pgvector
python-dateutil
